+++ zerver/lib/event_schema.py
@@ -1816,6 +1816,7 @@ group_type = DictType(
         ("direct_subgroup_ids", ListType(int)),
         ("description", str),
         ("is_system_group", bool),
+        ("can_manage_group", group_setting_type),
         ("can_mention_group", group_setting_type),
     ]
 )
@@ -1863,6 +1864,7 @@ user_group_data_type = DictType(
     optional_keys=[
         ("name", str),
         ("description", str),
+        ("can_manage_group", group_setting_type),
         ("can_mention_group", group_setting_type),
     ],
 )